+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Version 0.2.0
 
 * Camel case for class name
+* Nullable properties
+* Options for converter